Clean up!

+++ src/com/example/git/RepositoryDatabase.java
@@ -1,16 +1,21 @@
 package com.example.git;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import android.content.Context;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
 
 /**
- * This class handles the storage of the different 
+ * This class handles the storage of the different repositories.
  *
  */
-public class SqlLiteDatabaseHelper extends SQLiteOpenHelper {
+public final class RepositoryDatabase extends SQLiteOpenHelper {
 
-	private static SqlLiteDatabaseHelper mInstance;
+	private static RepositoryDatabase mInstance;
 	private static final String DATABASE_NAME = "woopwoop";
 	private static final int DATABASE_VERSION = 2;
 	private static final String TABLE_NAME = "Repositories";
@@ -20,7 +25,7 @@ public class SqlLiteDatabaseHelper extends SQLiteOpenHelper {
 	 * 
 	 * @param context
 	 */
-	private SqlLiteDatabaseHelper(Context context) {
+	private RepositoryDatabase(Context context) {
 		super(context, DATABASE_NAME, null, DATABASE_VERSION);
 	}
 
@@ -29,9 +34,9 @@ public class SqlLiteDatabaseHelper extends SQLiteOpenHelper {
 	 * @param context
 	 * @return
 	 */
-	public static SqlLiteDatabaseHelper getInstance(Context context) {
+	public synchronized static RepositoryDatabase getInstance(Context context) {
 		if (mInstance == null) {
-			mInstance = new SqlLiteDatabaseHelper(context.getApplicationContext());
+			mInstance = new RepositoryDatabase(context.getApplicationContext());
 		}
 		return mInstance;
 	}
@@ -41,7 +46,12 @@ public class SqlLiteDatabaseHelper extends SQLiteOpenHelper {
 	 * 
 	 */
 	public void onCreate(SQLiteDatabase db) {
-		db.execSQL(TABLE_CREATE);
+		if(db.isOpen()){
+			Log.d("database", "on create");
+			db.execSQL(TABLE_CREATE);
+		} else {
+			Log.d("database", "on create else");
+		}
 	}
 
 	@Override
@@ -57,17 +67,71 @@ public class SqlLiteDatabaseHelper extends SQLiteOpenHelper {
 	 * Inserts a given path to a Git repository into the database.
 	 * @param path The path to a Git repository that will be inserted.
 	 */
-	public void insertRepositoryPathintoTableRepositories(String path) {
+	public void addRepository(String path) {
 		SQLiteDatabase database = getWritableDatabase();
-		database.execSQL("INSERT INTO " + TABLE_NAME + " ('repoPath') VALUES ('" +  path + "');");
+		if (database.isOpen()) {
+			database.execSQL("INSERT INTO " + TABLE_NAME + " ('repoPath') VALUES ('" +  path + "');");
+			database.close();
+		} else {
+			Log.e("database", "Can't open database!");
+		}
 	}
 
 	/**
 	 * Removes a given path of a Git repository from the database. 
 	 * @param path The path to a Git repository that will be removed.
 	 */
-	public void removeRepositoryPathfromTableRepositories(String path) {
+	public void removeRepository(String path) {
 		SQLiteDatabase database = getWritableDatabase();
-		database.execSQL("DELETE FROM " + TABLE_NAME + " WHERE repoPath = '" + path + "';");
+		if (database.isOpen()) {
+			database.execSQL("DELETE FROM " + TABLE_NAME + " WHERE repoPath = '" + path + "';");
+			database.close();
+		} else {
+			Log.e("database", "Can't open database!");
+		}
+	}
+
+	/**
+	 * Loads all 
+	 * @return
+	 */
+	public List<String> loadRepositories() {
+		List<String> resultList = new ArrayList<String>();
+		SQLiteDatabase database = getWritableDatabase();
+		if (database.isOpen()) {
+			final String[] columns = new String[]{"repoPath"};
+			resultList = loadStringsFromDataBaseTable(database, TABLE_NAME, columns, " ");
+			database.close();
+		} else {
+			Log.d("database", "Can't open database!");
+		}
+		return resultList;
+	}
+
+	/**
+	 * Queries the data from specific columns of an SQLite database table and
+	 * writes them as string in the resultList.
+	 * @param database The database where the data be queried
+	 * @param table The table where the data will be queried
+	 * @param resultList The List where the queried data will be applied
+	 * @param columns The columns of the table where the data will be queried
+	 */
+	private List<String> loadStringsFromDataBaseTable(SQLiteDatabase database, String table, String[] columns, String spacer) {
+		List<String> resultList = new ArrayList<String>();
+		Cursor cursor = database.query(table, columns, null, null, null, null, null);
+		if (cursor != null) {
+			while(cursor.moveToNext()) {
+				String rowResult = "";
+				for (int i = 0; i < columns.length; i++) {
+					rowResult += new String(cursor.getString(0));
+					if(i != (columns.length - 1)) {
+						rowResult += spacer;
+					}
+				}
+				resultList.add(rowResult);
+			}
+			cursor.close();
+		}
+		return resultList;
 	}
 }
